Third Commit
This one contains working importation of animations from Blender as well as the system that captures the animation of the movement of the camera through the iPhone gyroscope. It successfully records animations through UI for the Camera transform and stuff. I think I have enough working to finally start working on transferring data between Blender and Unity.