Fourth Commit
This is the one that fully integrates data sending to Blender through a text file, lol. Parsing by JSON will probably come later through the project progression. Um, for the most part it's just getting the values correctly and formatting them correctly so it is easy to read and to send through a text file so that the values can them be plucked out fairly easily and put into the properties in Blender. Although there is still much needed work to create the finished product as I was just testing whether values can be sent through Unity to Blender and yes! The extraction of the values though probably requires more thourough reformatting of the values until it becomes efficient. This is practically just showing that yes we can create and set values in a timeline for animation and send them through a text file to Blender that can be read INSIDE Blender. I probably will just need to tweak a few more things in code to finish the C# and Unity side of things before comepletely starting the development of the Blender add-on. These are just purely to show that the movements and keyframes can be transported, and that the movement can be replicated; although I do see a problem with having the proper rotation of things, that will be an issue to fix in the future once I make sure that everything is working. Quaternion values have been truncated to the tenths place which I could have a problem with so I will probably fix that later once I make sure things work properly.