Fifth Commit (Working Project)
So this is the next big commit. This includes all of the working files and everything necessary outside of the needed software. So aside from Blender and Unity/Unity Remote 5, everything needed is here. Some things in Unity were changed around a tiny bit to make it more friendly o the python code. Speaking of, the python code for the add on is finally here and it works just as I had it. Only problem is that it's a one off for some reason so every time I use it and want to do it again i have to delete then reinstall the add-on. There is also a bunch of things I have to get fixed. But this commit is just to put out what is already working so I can refine it a bit later and stuff.